+++ rancher2/schema_cluster_v2_rke_config.go
@@ -122,6 +122,33 @@ func clusterV2RKEConfigFields() map[string]*schema.Schema {
 				Schema: clusterV2RKEConfigETCDFields(),
 			},
 		},
+		"rotate_certificates": {
+			Type:        schema.TypeList,
+			MaxItems:    1,
+			Optional:    true,
+			Description: "Cluster V2 certificate rotation",
+			Elem: &schema.Resource{
+				Schema: clusterV2RKEConfigRotateCertificatesFields(),
+			},
+		},
+		"etcd_snapshot_create": {
+			Type:        schema.TypeList,
+			MaxItems:    1,
+			Optional:    true,
+			Description: "Cluster V2 etcd snapshot create",
+			Elem: &schema.Resource{
+				Schema: clusterV2RKEConfigETCDSnapshotCreateFields(),
+			},
+		},
+		"etcd_snapshot_restore": {
+			Type:        schema.TypeList,
+			MaxItems:    1,
+			Optional:    true,
+			Description: "Cluster V2 etcd snapshot restore",
+			Elem: &schema.Resource{
+				Schema: clusterV2RKEConfigETCDSnapshotRestoreFields(),
+			},
+		},
 	}
 
 	return s