+++ rancher2/structure_cluster_v2_rke_config_z_test.go
@@ -34,6 +34,9 @@ func init() {
 	testClusterV2RKEConfigConf.MachineSelectorConfig = testClusterV2RKEConfigSystemConfigConf
 	testClusterV2RKEConfigConf.Registries = testClusterV2RKEConfigRegistryConf
 	testClusterV2RKEConfigConf.ETCD = testClusterV2RKEConfigETCDConf
+	testClusterV2RKEConfigConf.RotateCertificates = testClusterV2RKEConfigRotateCertificatesConf
+	testClusterV2RKEConfigConf.ETCDSnapshotCreate = testClusterV2RKEConfigETCDSnapshotCreateConf
+	testClusterV2RKEConfigConf.ETCDSnapshotRestore = testClusterV2RKEConfigETCDSnapshotRestoreConf
 
 	testClusterV2RKEConfigInterface = []interface{}{
 		map[string]interface{}{
@@ -45,6 +48,9 @@ func init() {
 			"machine_selector_config": testClusterV2RKEConfigSystemConfigInterface,
 			"registries":              testClusterV2RKEConfigRegistryInterface,
 			"etcd":                    testClusterV2RKEConfigETCDInterface,
+			"rotate_certificates":     testClusterV2RKEConfigRotateCertificatesInterface,
+			"etcd_snapshot_create":    testClusterV2RKEConfigETCDSnapshotCreateInterface,
+			"etcd_snapshot_restore":   testClusterV2RKEConfigETCDSnapshotRestoreInterface,
 		},
 	}
 }